First commit.
Current state of the webiste... We have a 'Post' form that takes 'name'
and 'text' as input.  These two inputs are used to create an instance of
the 'Post' class.  This new instance is added to the dummy 'test posts'
list.  No database created yet.
We have two routes, /home and /scratch.  /scratch is used for dummy
formatting.